calculate "frecency" in HistoryFilterModel
"Frecency" = "frequency" + "recency" is a useful metric for
sorting location bar autocompletion results. This change makes
the HistoryFilterModel calculate and update it for each history
entry that it tracks. The time bins and weighting factors used in
the calculation match the defaults in Firefox 3.0 -- but we do it
with integer math, not floating-point.