recalculate frecencies daily
Since the user may want to leave their browser open for more than
one day, recalculate frecencies daily so that older history entries
do have their frecency dropi over time. We don't need to do this
often, and it may take a second or two, so schedule it for 3am. If
the machine is suspended at that time, it'll just happen when the
machine is woken up. (If the machine is off, we'll recalculate when
the user restarts Arora, obviously.)