location bar auto-completion improvements
Hack around QCompleter's inflexibility, enabling us to find completions by
looking for the user's typed string as a substring of our history urls
and titles; also sort the results by frecency so that more interesting
hits are at the top of the list. (See historycompleter.h for the
dirty details of the implementation.) Also, update the autotests to not
expect HistoryCompletionModel to have twice as many rows as the
original HistoryFilterModel.