don't calculate the completion list while the user is typing
Updating the completion list can be time-consuming, especially if
we only have one or two letters to search for (and so get a lot
of hits). Instead, push off updating the completion list until
the user stops typing for a moment, and we hopefully have a more
restrictive search criterion.

To do this, we update HistoryCompletionModel with a notion of
validity. If it is invalid, it tells QCompleter that nothing matches.
We mark the model invalid when the user types something new, and then
update the model and mark it valid after they pause for a moment. If
the user is simply extending a preexisting search string, the new
results will be a strict subset of the current results, so we don't
invalidate the model, only schedule an update.