Revert "Add check for the Oxygen style and switch to plastique unless the user specifies oxygen on the command line"
This reverts commit 34a819786a82ead0bc56794520e7ca97f6663826.

One of the main reason we did the Oxygen style check was to stop all
the bug reports, but we still received a lot of bug reports, only this
time it is that we were not picking up the KDE style.  We will just
continue to point our users to the Oxygen developers and maybe they
will backport the Oxygen fixes to KDE 4.2.x.