Revert "Fix progress bar background with Oxygen style"
The patch introduced a regression affecting styles that use the
palette colour rather then the brush with our pixmap to paint the
background. And funnily enough the brush was considered to have a
black colour (not sure if it's a bug or a feature, might be worth
investigating), we ended up having the location bar painted in black.
Also, the patch was actually a workaround, a better long-term
solution would be to just fix it in Oxygen.

This reverts commit bb4085cd0182323a8e5062eb3228e3c6853336e4.