Add new ModelToolBar class, that is a subclass of QToolBar and an equivalent of ModelMenu for toolbars. It can be used to populate a toolbar with actions based on a given model.
It rebuilds itself on any change in the model (which can be improved).
For actions it uses to display role for text and decoration role for
icon. It uses ModelMenu for items that have children.

ModelToolBar supports drag and drop, not only copying but also moving.
With some changes in ModelMenu it will be also able to drag onto a
model menu and accept drops from it.

It install itself as an event filter for buttons so that it can
intercept mouse press and release events properly.