When the background of the window is not white hardcode the style of the webview to the commonstyle's palette.
When users are using a dark style a browser's background should not be
black even when the rest of the UI is dark.

This is only enabled on 4.6 as I noticed in Qt 4.5 the text is painted
from the style and not the palette and white on white just doesn't work.

Issue: 221