Fix Text Encoding to work correctly
Issue: 891

Store the name of the codec in the action so when the action is
activated the name can be correctly retrieved.  Storing the offset
requires that the same list be obtained which can introduce bugs such
as this one where the menu was enhanced to be sorted while the offset
was used against the unsorted list.